+++ cli/internal/cmd/runner_proc_unix.go
@@ -1,0 +1,24 @@
+//go:build !windows
+
+package cmd
+
+import (
+	"os/exec"
+	"syscall"
+)
+
+func hostExecSysProcAttr() *syscall.SysProcAttr {
+	return &syscall.SysProcAttr{Setpgid: true}
+}
+
+func killRunnerJobProcess(cmd *exec.Cmd) {
+	if cmd == nil || cmd.Process == nil {
+		return
+	}
+	pgid := cmd.Process.Pid
+	if cmd.SysProcAttr != nil && cmd.SysProcAttr.Setpgid {
+		_ = syscall.Kill(-pgid, syscall.SIGKILL)
+		return
+	}
+	_ = cmd.Process.Kill()
+}